fixed components issue
